fix: Use large data types for files sizes where appropriate.
Switch data types to use int64_t instead of `size_t`, which is only 32
bits on 32-bit platforms, in places where we expect to handle file
sizes.

Also enable error messages when we try to auto-convert from one type
to another in a lossy fashion.

Modified-by: Kristian Amlie <kristian.amlie@northern.tech>
The original contains additional compile warnings (which are treated
as errors). But since the cmake build files are structured differently
in older in branches, these end up being applied too broadly, hitting
the vendor code which is not prepared to pass their checks. So I
removed that part of the commit.

Changelog: Fix possible integer overflow when dealing with large files
on 32-bit platforms.

Ticket: MEN-7613

(cherry picked from commit 14ec89ef374ec78a2e8c10a4861888d114804443)

+++ tests/src/mender-update/deployments_test.cpp
@@ -932,7 +932,8 @@ TEST_F(DeploymentsTests, JsonLogMessageReaderTest) {
 	EXPECT_EQ(deps::JsonLogMessagesReader::TotalDataSize(messages.size() - 1), expected_total_size);
 
 	auto file_reader = make_shared<io::FileReader>(test_log_file_path);
-	deps::JsonLogMessagesReader logs_reader {file_reader, messages.size() - 1};
+	deps::JsonLogMessagesReader logs_reader {
+		file_reader, static_cast<int64_t>(messages.size() - 1)};
 
 	stringstream ss;
 	vector<uint8_t> buf(1024);
@@ -971,7 +972,8 @@ TEST_F(DeploymentsTests, JsonLogMessageReaderSmallBufferTest) {
 	EXPECT_EQ(deps::JsonLogMessagesReader::TotalDataSize(messages.size() - 1), expected_total_size);
 
 	auto file_reader = make_shared<io::FileReader>(test_log_file_path);
-	deps::JsonLogMessagesReader logs_reader {file_reader, messages.size() - 1};
+	deps::JsonLogMessagesReader logs_reader {
+		file_reader, static_cast<int64_t>(messages.size() - 1)};
 
 	stringstream ss;
 	vector<uint8_t> buf(16);
@@ -1010,7 +1012,8 @@ TEST_F(DeploymentsTests, JsonLogMessageReaderSmallEvenBufferTest) {
 	EXPECT_EQ(deps::JsonLogMessagesReader::TotalDataSize(messages.size() - 1), expected_total_size);
 
 	auto file_reader = make_shared<io::FileReader>(test_log_file_path);
-	deps::JsonLogMessagesReader logs_reader {file_reader, messages.size() - 1};
+	deps::JsonLogMessagesReader logs_reader {
+		file_reader, static_cast<int64_t>(messages.size() - 1)};
 
 	stringstream ss;
 	vector<uint8_t> buf(7);
@@ -1049,7 +1052,8 @@ TEST_F(DeploymentsTests, JsonLogMessageReaderRewindTest) {
 	EXPECT_EQ(deps::JsonLogMessagesReader::TotalDataSize(messages.size() - 1), expected_total_size);
 
 	auto file_reader = make_shared<io::FileReader>(test_log_file_path);
-	deps::JsonLogMessagesReader logs_reader {file_reader, messages.size() - 1};
+	deps::JsonLogMessagesReader logs_reader {
+		file_reader, static_cast<int64_t>(messages.size() - 1)};
 
 	stringstream ss;
 	vector<uint8_t> buf(1024);
